requirements.txt
anyio==1.3.0
APScheduler==2.1.2
cachetools==4.2.2
certifi==2019.11.28
click==6.7
cmake==3.18.4.post1
dlib==19.7.0
face-recognition==1.2.3
face-recognition-models==0.3.0
h11==0.8.1
httpcore==0.12.3
httpx==0.16.1
idna==2.10
numpy==1.16.6
opencv-python==3.4.14.0
Pillow==6.2.2
python-dotenv==0.10.3
python-telegram-bot==12.8
pytz==2020.5
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.3.2
six==1.12.0
sniffio==1.1.0
tornado==5.1.1
tzdata==2020.1
tzlocal==1.5.1
